requirements.txt
langchain-openai 
langgraph
streamlit 
tavily-python
python-dotenv